TheLight24/requirements.txt
numpy
matplotlib
pybind11
pygame==2.5.2
sounddevice
pyaudio
vosk==0.3.45
pvporcupine
openai
python-dotenv
torch==2.4.0+cpu ; sys_platform == 'linux'
TTS==0.22.0
webrtcvad
opencv-python==4.10.0.84
pillow
pytest
